Make code more horizon-like ("horizontal"??).
There are no longer any api calls in tabs.py: the Tabs and Tables only
know about the data that gets passed to them from on high (views.py),
which feels a bit cleaner.
